+++ aristotle_mdr/contrib/autocomplete/views.py
@@ -43,7 +43,7 @@ class GenericAutocomplete(autocomplete.Select2QuerySetView):
         """Return the label of a result."""
 
         template = get_template(self.template_name)
-        context = Context({"result": result})
+        context = Context({"result": result, 'request': self.request})
         return template.render(context)
 
     def get_results(self, context):
@@ -92,10 +92,10 @@ class GenericConceptAutocomplete(GenericAutocomplete):
         ]
 
 
-class UserAutocomplete(autocomplete.Select2QuerySetView):
+class UserAutocomplete(GenericAutocomplete):
     # model = User
     model = None
-    template_name = "autocomplete_light/item.html"
+    template_name = "aristotle_mdr/actions/autocompleteUser.html"
 
     def dispatch(self, request, *args, **kwargs):
         return super(UserAutocomplete, self).dispatch(request, *args, **kwargs)
@@ -111,18 +111,37 @@ class UserAutocomplete(autocomplete.Select2QuerySetView):
             raise PermissionDenied
 
         if self.q:
-            if self.request.user.is_superuser:
-                qs = self.model.objects.filter(
-                    Q(username__icontains=self.q) | Q(email__icontains=self.q)
-                )
-            else:
-                qs = self.model.objects.filter(
-                    Q(username__iexact=self.q) | Q(email__iexact=self.q)
-                )
+            qs = self.model.objects.filter(is_active=True).filter(
+                Q(username__icontains=self.q) |
+                Q(email__icontains=self.q) |
+                Q(first_name__icontains=self.q) | Q(last_name__icontains=self.q)
+            )
         else:
             if self.request.user.is_superuser:
-                qs = self.model.objects.all()
+                qs = self.model.objects.filter(is_active=True)
             else:
                 qs = self.model.objects.none()
 
         return qs
+
+    def get_result_text(self, result):
+        """Return the label of a result."""
+
+        template = get_template(self.template_name)
+        result.highlight = {}
+        for f in ['username', 'email', 'get_full_name']:
+            field = getattr(result, f, None)
+            if callable(field):
+                field = field()
+            if field and self.q.lower() in field.lower():
+                index = field.lower().index(self.q.lower())
+                offset = index + len(self.q)
+                field = "%s<b><u>%s</u></b>%s" % (field[:index], field[index:offset], field[offset:])
+
+            result.highlight[f] = field
+        context = Context({"result": result, 'request': self.request})
+        return template.render(context)
+
+    def get_result_title(self, result):
+        """Return the title of a result."""
+        return six.text_type(result)