+++ aristotle_mdr/forms/workgroups.py
@@ -4,6 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 
 import aristotle_mdr.models as MDR
 from aristotle_mdr.contrib.autocomplete import widgets
+from aristotle_mdr.forms.creation_wizards import UserAwareForm
 
 
 class AddMembers(forms.Form):
@@ -22,3 +23,12 @@ class AddMembers(forms.Form):
         roles = self.cleaned_data['roles']
         roles = [role for role in roles if role in MDR.Workgroup.roles.keys()]
         return roles
+
+
+class ChangeWorkgroupUserRolesForm(UserAwareForm):
+    roles = forms.MultipleChoiceField(
+        label=_("Workgroup roles"),
+        choices=sorted(MDR.Workgroup.roles.items()),
+        widget=forms.CheckboxSelectMultiple,
+        required=False
+    )