+++ aristotle_mdr/required_settings.py
@@ -24,8 +24,13 @@ DATABASES = {
 
 CACHES = {
     'default': {
-        'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
-        'LOCATION': 'aristotle-mdr-cache'
+        'BACKEND': 'django.core.cache.backends.filebased.FileBasedCache',
+        'LOCATION': os.path.join(BASE_DIR, 'caches', 'aristotle-mdr-cache'),
+    },
+    'aristotle-mdr-invitations': {
+        'BACKEND': 'django.core.cache.backends.filebased.FileBasedCache',
+        'LOCATION': os.path.join(BASE_DIR, 'caches', 'aristotle-mdr-invitations'),
+        'TIMEOUT': 60 * 60 * 24 * 7,  # sec * min * hours * days
     }
 }
 
@@ -78,6 +83,7 @@ INSTALLED_APPS = (
     'aristotle_mdr.contrib.slots',
     'aristotle_mdr.contrib.identifiers',
     'aristotle_mdr.contrib.browse',
+    'aristotle_mdr.contrib.user_management',
 
     'channels',
     'haystack_channels',
@@ -105,6 +111,7 @@ INSTALLED_APPS = (
     'reversion_compare',  # https://github.com/jedie/django-reversion-compare
 
     'notifications',
+    'organizations',
 )
 
 USE_L10N = True
@@ -195,7 +202,14 @@ ARISTOTLE_SETTINGS = {
         # (fileType, menu, font-awesome-icon, module)
         # ('csv-vd', 'CSV list of values', 'fa-file-excel-o', 'aristotle_mdr', 'CSV downloads for value domain codelists'),
         'aristotle_mdr.downloader.CSVDownloader'
-    ]
+    ],
+
+    # These settings aren't active yet.
+    # "USER_EMAIL_RESTRICTIONS": None,
+    "USER_VISIBILITY": ['owner', 'workgroup_manager', 'registation_authority_manager']
+    # "SIGNUP_OPTION": 'closed', # or 'closed'
+    # "GROUPS_CAN_INVITE": 'closed', # or 'closed'
+
 }
 
 CKEDITOR_CONFIGS = {
@@ -226,3 +240,5 @@ HAYSTACK_CONNECTIONS = {
 STATIC_PRECOMPILER_COMPILERS = (
     ('static_precompiler.compilers.LESS', {"executable": "lesscpy"}),
 )
+
+ORGS_SLUGFIELD = 'autoslug.fields.AutoSlugField'