+++ aristotle_mdr/urls/__init__.py
@@ -9,6 +9,7 @@ admin.autodiscover()
 
 urlpatterns = [
     url(r'^', include('aristotle_mdr.urls.base')),
+    url(r'^', include('aristotle_mdr.contrib.user_management.urls', namespace="aristotle-user")),
     url(r'^', include('aristotle_mdr.urls.aristotle', app_name="aristotle_mdr", namespace="aristotle")),
     url(r'^ac/', include('aristotle_mdr.contrib.autocomplete.urls', namespace="aristotle-autocomplete")),
     url(r'^', include('aristotle_mdr.contrib.healthcheck.urls', app_name="aristotle_mdr_hb", namespace="aristotle_hb")),