+++ aristotle_mdr/views/utils.py
@@ -1,13 +1,22 @@
-from braces.views import PermissionRequiredMixin
+from braces.views import LoginRequiredMixin, PermissionRequiredMixin
+
+from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.core.urlresolvers import reverse
 from django.db.models import Count, Q
-from django.shortcuts import render
-from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
 from django.db.models.functions import Lower
+from django.http import Http404
+from django.shortcuts import redirect, render, get_object_or_404
+from django.utils import timezone
+from django.utils.functional import cached_property
+from django.utils.translation import ugettext_lazy as _
+
+from django.views.generic.detail import BaseDetailView
+from django.views.generic import (
+    CreateView, DetailView, FormView, ListView, RedirectView, UpdateView
+)
 
 
 paginate_sort_opts = {
@@ -179,23 +188,6 @@ def paginated_registration_authority_list(request, ras, template, extra_context=
     return render(request, template, context)
 
 
-def get_concept_redirect_or_404(get_item_perm, request, iid, objtype=None):
-    if objtype is None:
-        from aristotle_mdr.models import _concept
-        objtype = _concept
-
-    from aristotle_mdr.contrib.redirect.exceptions import Redirect
-
-    item = get_item_perm(objtype, request.user, iid)
-    if not item:
-        if request.user.is_anonymous():
-            raise Redirect(reverse('friendly_login') + '?next=%s' % request.path)
-        else:
-            raise PermissionDenied
-    else:
-        return item
-
-
 def workgroup_item_statuses(workgroup):
     from aristotle_mdr.models import STATES
 
@@ -246,3 +238,59 @@ class ObjectLevelPermissionRequiredMixin(PermissionRequiredMixin):
         else:
             has_permission = request.user.has_perm(self.get_permission_required(request))
         return has_permission
+
+
+class GroupMemberMixin(object):
+    user_pk_kwarg = "user_pk"
+
+    @cached_property
+    def user_to_change(self):
+        user = get_object_or_404(get_user_model(), pk=self.kwargs.get(self.user_pk_kwarg))
+        if user not in self.get_object().members.all():
+            raise Http404
+        return user
+
+    def get_context_data(self, **kwargs):
+        """
+        Insert the single object into the context dict.
+        """
+        kwargs = super(GroupMemberMixin, self).get_context_data(**kwargs)
+        kwargs.update({'user_to_change': self.user_to_change})
+        return kwargs
+
+
+class RoleChangeView(GroupMemberMixin, LoginRequiredMixin, ObjectLevelPermissionRequiredMixin, BaseDetailView, FormView):
+    raise_exception = True
+    redirect_unauthenticated_users = True
+    object_level_permissions = True
+
+    def get_form_kwargs(self):
+        kwargs = super(RoleChangeView, self).get_form_kwargs()
+        kwargs.update({'user': self.request.user})
+        initial = {'roles': []}
+        initial['roles'] = self.get_object().list_roles_for_user(self.user_to_change)
+
+        kwargs.update({'initial': initial})
+        return kwargs
+
+    def form_valid(self, form):
+        for role in self.model.roles:
+            if role in form.cleaned_data['roles']:
+                self.get_object().giveRoleToUser(role, self.user_to_change)
+            else:
+                self.get_object().removeRoleFromUser(role, self.user_to_change)
+
+        return self.get_success_url()
+
+
+class MemberRemoveFromGroupView(GroupMemberMixin, LoginRequiredMixin, ObjectLevelPermissionRequiredMixin, DetailView):
+    raise_exception = True
+    redirect_unauthenticated_users = True
+    object_level_permissions = True
+
+    http_method_names = ['get', 'post']
+
+    def post(self, request, *args, **kwargs):
+        for role in self.get_object().list_roles_for_user(self.user_to_change):
+            self.get_object().removeRoleFromUser(role, self.user_to_change)
+        return self.get_success_url()